pyproject.toml
[tool.poetry]
name = "hospital-management-system"
version = "0.1.0"
description = ""
authors = ["Subhomoy Roy Choudhury"]
license = "MIT"
readme = "README.md"
packages = [{include = "main"}]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.15.1"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
